+++ src/v8serialize/__init__.py
@@ -1,5 +1,11 @@
 from __future__ import annotations
 
+from v8serialize._errors import DecodeV8CodecError as DecodeV8CodecError
+from v8serialize._errors import JSRegExpV8CodecError as JSRegExpV8CodecError
+from v8serialize._errors import (
+    UnmappedTagDecodeV8CodecError as UnmappedTagDecodeV8CodecError,
+)
+from v8serialize._errors import V8CodecError as V8CodecError
 from v8serialize.constants import SerializationFeature as SerializationFeature
 from v8serialize.decode import default_tag_mappers as default_tag_mappers
 from v8serialize.decode import loads as loads
@@ -12,9 +18,3 @@ from v8serialize.encode import (
 )
 from v8serialize.encode import default_object_mappers as default_object_mappers
 from v8serialize.encode import dumps as dumps
-from v8serialize.errors import DecodeV8CodecError as DecodeV8CodecError
-from v8serialize.errors import JSRegExpV8CodecError as JSRegExpV8CodecError
-from v8serialize.errors import (
-    UnmappedTagDecodeV8CodecError as UnmappedTagDecodeV8CodecError,
-)
-from v8serialize.errors import V8CodecError as V8CodecError
