+++ src/v8serialize/_errors.py
