+++ src/v8serialize/extensions.py
@@ -6,9 +6,9 @@ from enum import Enum
 from typing import TYPE_CHECKING
 
 from v8serialize._enums import frozen
+from v8serialize._errors import DecodeV8CodecError
 from v8serialize._pycompat.dataclasses import slots_if310
 from v8serialize.constants import ArrayBufferViewTag
-from v8serialize.errors import DecodeV8CodecError
 from v8serialize.jstypes.jsbuffers import (
     ArrayBufferViewStructFormat,
     JSArrayBuffer,
