+++ src/v8serialize/jstypes/jsobject.py
@@ -7,8 +7,8 @@ from itertools import chain
 from types import MappingProxyType
 from typing import TYPE_CHECKING, Any, overload
 
+from v8serialize._errors import NormalizedKeyError
 from v8serialize._pycompat.dataclasses import slots_if310
-from v8serialize.errors import NormalizedKeyError
 from v8serialize.jstypes import _repr
 from v8serialize.jstypes._normalise_property_key import normalise_property_key
 from v8serialize.jstypes.jsarrayproperties import (
