+++ src/v8serialize/jstypes/jsregexp.py
@@ -5,10 +5,10 @@ from dataclasses import dataclass, field
 from re import Pattern, compile
 from typing import AnyStr, Literal, overload
 
+from v8serialize._errors import JSRegExpV8CodecError
 from v8serialize._pycompat.dataclasses import slots_if310
 from v8serialize._pycompat.re import RegexFlag
 from v8serialize.constants import JSRegExpFlag
-from v8serialize.errors import JSRegExpV8CodecError
 
 
 @dataclass(frozen=True, order=True, **slots_if310())
